Restructure, add multiprocessing, viz with plt.Triangulation
Added multiprocessing using python's Pool to parallellize sampling
photons and calculating different isoradials. Significant speedup.

Resturcture to allow for package publication later on

WIP: visualization using matplotlib's triangulation library is very
neat. Working on this
