+++ .gitignore
@@ -1,0 +1,4 @@
+
+# pixi environments
+.pixi
+*.egg-info
