--- Tests/parameters.ini
@@ -1,41 +1,0 @@
-[isoradial_angular_parameters]
-start_angle             =   0
-end_angle               =   np.pi
-angular_precision       =   50
-mirror                  =   True  # if Ture, calculates only half of the isoradial and mirrors it
-
-[solver_parameters]  # parameters for solving equation 13
-initial_guesses             =   12
-midpoint_iterations         =   11
-plot_inbetween              =   False  # plot isoredshifts while improving them
-min_periastron              =   3.001  # minimum distance to black hole, in units of black hole mass (photon sphere is at 3M)
-use_ellipse                 =   True
-
-[isoredshift_solver_parameters]
-initial_guesses             =   12
-midpoint_iterations         =   10
-times_inbetween             =   2   # amount of times to double the precision of an isoredshift line when improving
-retry_angular_precision     =   15  # angular precision to calculate isoradials with when improving solutions
-plot_inbetween              =   False  # plot isoredshifts while improving them
-min_periastron              =   3.01  # minimum distance to black hole (must be strictly larger than 3M), in units of black hole mass (photon sphere is at 3M)
-use_ellipse                 =   True
-retry_tip                   =   10
-initial_radial_precision    =   10
-
-[plot_params]
-plot_isoredshifts_inbetween =   False
-save_plot                   =   False
-plot_ellipse                =   False
-plot_core                   =   True
-redshift                    =   True
-linestyle                   =   "-"
-key                         =   ""
-face_color                  =   "black"
-line_color                  =   "white"
-text_color                  =   "white"
-alpha                       =   1.
-show_grid                   =   False
-legend                      =   False
-orig_background             =   False
-plot_disk_edges             =   False
-ax_lim                      =   (-35, 35)
