+++ luminet/__init__.py
@@ -1,0 +1,3 @@
+"""
+    Calculate and plot Swarzschild black holes with a thin accretion disk
+"""
